+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+- Support SPAs by blocking links incrementally via `MutationObserver`.
+
 ## [0.2.1] - 2019-09-16
 
 ### Fixed